Move loss aggregation from svir.py to aggregate_loss_by_zone.py
